+++ package.json
@@ -18,6 +18,7 @@
         "cors": "^2.8.5",
         "dotenv": "^16.3.1",
         "express": "^4.18.2",
+        "express-rate-limit": "^6.7.1",
         "jsonwebtoken": "^9.0.1",
         "jwt-simple": "^0.5.6",
         "moment": "^2.29.4",
