+++ src/middlewares/index.js
@@ -1,0 +1,7 @@
+const productsUploads = require('./multerStorage');
+const rateLimiter = require('./rateLimiter');
+
+module.exports = {
+    productsUploads,
+    rateLimiter
+};
