+++ src/middlewares/rateLimiter.js
@@ -1,0 +1,9 @@
+const RateLimiter = require('express-rate-limit');
+
+const rateLimiter = new RateLimiter({
+    windowMs: 1000 * 60 * 60,
+    max: 100,
+    message: 'Excediste el número de peticiones por hora'
+});
+
+module.exports = rateLimiter;
