+++ src/routes/index.js
@@ -1,4 +1,3 @@
-const multer = require("multer");
 const {
     getAllProductsHandler,
     getCategoriesHandler,
@@ -9,30 +8,15 @@ const {
 } = require('../handlers/');
 const { Router } = require('express');
 const router = Router();
+const { productsUploads, rateLimiter } = require('../middlewares/');
 
 
-const storage = multer.diskStorage({
-
-    destination: (req, res, cb) => {
-
-        cb(null, "./src/assets/products");
-
-    },
-    filename: (req, file, cb) => {
-
-        cb(null, "product-" + Date.now() + "-" + file.originalname);
-    }
-});
-
-const productsUploads = multer({ storage });
-
-
+router.use(rateLimiter);
 router.get('/products', getAllProductsHandler);
 router.get('/categories', getCategoriesHandler);
 router.get('/details/:id', getProductByIdHandler);
 router.put("/products/:id", [productsUploads.single("image")],
     editProductHandler);
-//image es el nombre del field html por el cual se recibe el archivo;
 router.post('/products', [productsUploads.single("image")],
     createProductHandler);
 router.delete("/products/:id", deleteProductHandler);
